# Move series fetches out of inner loop of SortedPostings.