@krasi-georgiev 2018-12-28
@brian-brazil 	yeah some comments would be helpful.


Here are the results from the promql bench, but these aren't really helpful as after looking at the code it works only with very small number of series and this optimisation addresses sorting of large number of series.

```
benchcmp -mag -best oldSortedPostings.txt newSortedPostings.txt
benchmark                                                                                                              old ns/op      new ns/op      delta
BenchmarkRangeQuery/expr=a_hundred,steps=1000-48                                                                       16014186       13162284       -17.81%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=1-48                                                                91275236       103719663      +13.63%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=1-48                                                             107630486      96568275       -10.28%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=10000-48                                                                27481825       25142839       -8.51%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=1-48                                                  10200160       10814819       +6.03%
BenchmarkRangeQuery/expr=-a_ten,steps=1000-48                                                                          1415394        1335386        -5.65%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=10000-48                                                            271139618      256254422      -5.49%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=1000-48                                                                 2451494        2324181        -5.19%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=10-48                                                                   1212438        1278135        +5.42%
```