@brian-brazil 2018-12-28
Updated.

That looks like noise more than anything.